// src/modules/store/types/indent.types.ts
export interface IndentRecordDTO {
  id: number;
  indent_number: string;
  indenter_name: string;
  product_name: string;
  quantity: number;
  uom: string;
  attachment: string;
  specifications: string;
  area_of_use: string;
  vendor_type: string;
  indent_status: string;
  indent_type: string;
  no_day: number;
  planned1: string;
  actual1: string;
  firm_name: string;
  firm_id?: number;
  firmName: string;
  firmNameMatch: string;
  approved_quantity: number;
  timestamp: string;
  price: number;
  total_rate: number;
  indent_approved_by: string;
  approved_date: string;
  planned2: string;
  actual2: string;
  vendor_name: string;
  negotiated_rate: number;
  planned3: string;
  actual3: string;
  attachment3: string;
  comparative_analysis: string;
  planned4: string;
  actual4: string;
  po_number: string;
  po_date: string;
  po_copy: string;
  planned5: string;
  actual5: string;
  transportation_include: string;
  tax_extra: string;
  credit_days: number;
  remarks5: string;
  status: string;
  lifting_status?: string;
  po_qty?: number;
  received_quantity?: number;
  pending_qty?: number;
  approved_vendor_name?: string;
  pending_po_qty?: number | null;
  pending_lift_qty?: number | null;
  po_requred?: string;
  vendor1_rank?: string;
  vendor2_rank?: string;
  vendor3_rank?: string;
  indent_url?: string;
  expected_req_date?: string;
  group_head?: string;
  min_stock_qty?: number;
}